requirements.txt
fastapi
uvicorn
python-jose
passlib
pymongo
python-multipart
bcrypt
Jinja2
email-validator